pyproject.toml
[project]
name = "leverx-courses"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.10",
    "factory-boy>=3.3.3",
    "httpie>=3.2.4",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]
